+++ src/blocks/place-card/__delete-icon/place-card__delete-icon.css
