+++ src/blocks/places-list/places-list.css
