+++ src/blocks/popup/__content/popup__content_picture/popup__content_picture.css
