+++ src/blocks/popup/__picture/popup__picture_large/popup__picture_large.css
