+++ src/blocks/user-info/__photo/user-info__photo.css
