+++ src/index.js
@@ -1,0 +1,4 @@
+import "./script/script.js";
+import "./pages/index.css";
+
+const serverUrl = NODE_ENV === 'development' ? 'http://praktikum.tk/cohort4' : 'https://praktikum.tk/cohort4';
